movies/pyproject.toml
[project]
name = "movies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.2",
    "httpx>=0.28.1",
]
